+++ p3_datawarehouse_aws_redshift/p3src/createenv/sql_queries_create.py
