+++ p3_datawarehouse_aws_redshift/p3src/etl/__init__.py
@@ -1,0 +1,1 @@
+from p3_datawarehouse_aws_redshift.p3src.etl.etl import etl_main
