+++ p3_datawarehouse_aws_redshift/p3src/utils/__init__.py
@@ -1,0 +1,1 @@
+from p3_datawarehouse_aws_redshift.p3src.utils.aws import get_myip, get_cluster_properties
