+++ p3_datawarehouse_aws_redshift/p3src/utils/aws.py
