+++ p3_datawarehouse_aws_redshift/project/Readme.md
@@ -1,1 +1,7 @@
-TODO
+# Project Folder
+## Purpose
+- Store the configuration variables
+- Store the routines the can be called by the end-user or automated
+
+## Use
+- Make sure the module p3_datawarehouse_aws_
