+++ p3_datawarehouse_aws_redshift/project/routines/Readme.md
@@ -1,1 +1,11 @@
-TODO
+# Routines
+## Purpose
+* This routines can be called on and automated
+* They do:
+    * Create the cluster
+    * Create the tables
+    * ETL the data from S3 to the tables
+
+## Parameters
+* They need as input the path to the config file for the admin and the database user
+* Normally stored in the config file of the project
