+++ p3_datawarehouse_aws_redshift/project/routines/create_tables_main.py
@@ -1,4 +1,4 @@
-from p3_datawarehouse_aws_redshift.src import create_tables_main
+from p3_datawarehouse_aws_redshift.p3src import create_tables_main
 import configparser
 # Input here the path to the admin config file
 config_path = '/Users/paulogier/81-GithubPackages/Udacity-Data-Engineer-NanoDegree/p3_datawarehouse_aws_redshift/project/config/dbuser_config.cfg'
