+++ p3_datawarehouse_aws_redshift/project/routines/etl_main.py
@@ -1,4 +1,4 @@
-from p3_datawarehouse_aws_redshift.src import etl_main
+from p3_datawarehouse_aws_redshift.p3src import etl_main
 import configparser
 config_path = '/Users/paulogier/81-GithubPackages/Udacity-Data-Engineer-NanoDegree/p3_datawarehouse_aws_redshift/project/config/dbuser_config.cfg'
