--- p3_datawarehouse_aws_redshift/src/__init__.py
@@ -1,3 +1,0 @@
-from p3_datawarehouse_aws_redshift.src.createenv import create_tables_main
-from p3_datawarehouse_aws_redshift.src.createenv import create_cluster_main
-from p3_datawarehouse_aws_redshift.src.etl import etl_main
