--- p3_datawarehouse_aws_redshift/src/createenv/__init__.py
@@ -1,2 +1,0 @@
-from p3_datawarehouse_aws_redshift.src.createenv.create_cluster import create_cluster_main
-from p3_datawarehouse_aws_redshift.src.createenv.create_tables import create_tables_main
