--- p3_datawarehouse_aws_redshift/src/etl/__init__.py
@@ -1,1 +1,0 @@
-from p3_datawarehouse_aws_redshift.src.etl.etl import etl_main
