--- p3_datawarehouse_aws_redshift/src/utils/__init__.py
@@ -1,1 +1,0 @@
-from p3_datawarehouse_aws_redshift.src.utils.aws import get_myip, get_cluster_properties
